postgres: Adjust head
The  traditional default  of  `import <nixpkgs>  {}`
didn't make sense as the Nix expression body is only
able  to  deal with  a  string  (a Git  SHA  commit,
to  be  specific),  and   not  a  package  set.  See
`baseline/baseline_config.nix`;  it does  retain the
seemingly contradictory default as it also has extra
code to distinguish whether the input is a string or
a package set (i.e., attribute set).