_UTILS.NIX: remove relPathToRelPathString [!!]
Fell back on the same  issue that Nix relative paths
cannot be used as input, because they will always be
interpreted from  the file  system position  the Nix
expression is called from if hardcoded, and there is
no reliable  way to pass  it (or it's just  too much
hassle), so  falling back to using  strings. Fingers
crossed.