POSTGRES: remove clean up step of deleting `postgres` (+ PGDATA) [..]
Changed  my  mind  deleting the  `postgres`  working
dir  after  exiting  the  Nix  shell.  If  used  for
development,  this  gives   more  control  over  the
environment -  and headaches as well  if one forgets
that there is such a dir  that will be reused on the
next run... Anyway, at the  moment, I see more value
in  leaving it  there and  deleting it  when testing
from a new backup, etc.

Also left a TODO, just in case.